46134-d-course-967/requirements/prod.txt
Django==4.2.5
django-cleanup==8.0.0
django-environ==0.11.2
django-tinymce==3.6.1
django-translation==1.2
Pillow==10.1.0
sorl-thumbnail==12.10.0
transliterate==1.10.2